+++ tests/test_plotting.py
@@ -313,6 +313,12 @@ class TestSinglePlotter(unittest.TestCase):
 class TestSinglePlotter1D(unittest.TestCase):
     def setUp(self):
         self.plotter = plotting.SinglePlotter1D()
+        self.dataset = aspecd.dataset.ExperimentalDataset()
+        self.dataset.data.data = np.random.random(5)
+        self.dataset.data.axes[0].quantity = 'magnetic field'
+        self.dataset.data.axes[0].unit = 'mT'
+        self.dataset.data.axes[1].quantity = 'intensity'
+        self.dataset.data.axes[1].unit = 'V'
 
     def tearDown(self):
         if self.plotter.fig:
@@ -417,6 +423,25 @@ class TestSinglePlotter1D(unittest.TestCase):
                          plotter.axes.get_xlim()[0])
         self.assertEqual(dataset_.data.data.min(),
                          plotter.axes.get_ylim()[0])
+
+    def test_has_switch_axes_parameter(self):
+        self.assertTrue('switch_axes' in self.plotter.parameters)
+
+    def test_switch_axes_sets_correct_axes_labels(self):
+        self.plotter.parameters['switch_axes'] = True
+        self.plotter.dataset = self.dataset
+        self.plotter.plot()
+        self.assertIn('intensity', self.plotter.ax.get_xlabel())
+        self.assertIn('magnetic', self.plotter.ax.get_ylabel())
+
+    def test_switch_axes_actually_switches_axes(self):
+        self.plotter.parameters['switch_axes'] = True
+        self.plotter.dataset = self.dataset
+        self.plotter.plot()
+        self.assertListEqual(
+            list(self.dataset.data.data),
+            list(self.plotter.axes.lines[0].get_xdata())
+        )
 
 
 class TestSinglePlotter2D(unittest.TestCase):
@@ -1064,6 +1089,12 @@ class TestMultiPlotter(unittest.TestCase):
 class TestMultiPlotter1D(unittest.TestCase):
     def setUp(self):
         self.plotter = plotting.MultiPlotter1D()
+        self.dataset = aspecd.dataset.ExperimentalDataset()
+        self.dataset.data.data = np.random.random(5)
+        self.dataset.data.axes[0].quantity = 'magnetic field'
+        self.dataset.data.axes[0].unit = 'mT'
+        self.dataset.data.axes[1].quantity = 'intensity'
+        self.dataset.data.axes[1].unit = 'V'
 
     def tearDown(self):
         if self.plotter.fig:
@@ -1130,6 +1161,77 @@ class TestMultiPlotter1D(unittest.TestCase):
         self.plotter.plot()
         self.assertEqual(dataset_.label,
                          self.plotter.legend.get_texts()[0].get_text())
+
+    def test_has_switch_axes_parameter(self):
+        self.assertTrue('switch_axes' in self.plotter.parameters)
+
+    def test_switch_axes_sets_correct_axes_labels(self):
+        self.plotter.parameters['switch_axes'] = True
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertIn('intensity', self.plotter.ax.get_xlabel())
+        self.assertIn('magnetic', self.plotter.ax.get_ylabel())
+
+    def test_switch_axes_actually_switches_axes(self):
+        self.plotter.parameters['switch_axes'] = True
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertListEqual(
+            list(self.dataset.data.data),
+            list(self.plotter.axes.lines[0].get_xdata())
+        )
+
+    def test_has_tight_parameter(self):
+        self.assertTrue('tight' in self.plotter.parameters)
+
+    def test_tight_sets_correct_x_axes_limits(self):
+        self.plotter.parameters['tight'] = 'x'
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertListEqual([0, 4], list(self.plotter.ax.get_xlim()))
+
+    def test_tight_sets_correct_y_axes_limits(self):
+        self.plotter.parameters['tight'] = 'y'
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertListEqual([min(self.dataset.data.data),
+                              max(self.dataset.data.data)],
+                             list(self.plotter.ax.get_ylim()))
+
+    def test_tight_sets_correct_axes_limits_with(self):
+        self.plotter.parameters['tight'] = 'both'
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertListEqual([0, 4], list(self.plotter.ax.get_xlim()))
+        self.assertListEqual([min(self.dataset.data.data),
+                              max(self.dataset.data.data)],
+                             list(self.plotter.ax.get_ylim()))
+
+    def test_tight_sets_correct_x_axes_limits_with_switched_axes(self):
+        self.plotter.parameters['switch_axes'] = True
+        self.plotter.parameters['tight'] = 'x'
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertListEqual([min(self.dataset.data.data),
+                              max(self.dataset.data.data)],
+                             list(self.plotter.ax.get_xlim()))
+
+    def test_tight_sets_correct_y_axes_limits_with_switched_axes(self):
+        self.plotter.parameters['switch_axes'] = True
+        self.plotter.parameters['tight'] = 'y'
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertListEqual([0, 4], list(self.plotter.ax.get_ylim()))
+
+    def test_tight_sets_correct_axes_limits_with_switched_axes(self):
+        self.plotter.parameters['switch_axes'] = True
+        self.plotter.parameters['tight'] = 'both'
+        self.plotter.datasets.append(self.dataset)
+        self.plotter.plot()
+        self.assertListEqual([min(self.dataset.data.data),
+                              max(self.dataset.data.data)],
+                             list(self.plotter.ax.get_xlim()))
+        self.assertListEqual([0, 4], list(self.plotter.ax.get_ylim()))
 
 
 class TestMultiPlotter1DStacked(unittest.TestCase):
@@ -1730,7 +1832,7 @@ class TestFigureProperties(unittest.TestCase):
         plt.close(plot.figure)
 
 
-class TestAxisProperties(unittest.TestCase):
+class TestAxesProperties(unittest.TestCase):
     def setUp(self):
         self.axis_properties = plotting.AxesProperties()
 
@@ -1748,7 +1850,9 @@ class TestAxisProperties(unittest.TestCase):
     def test_has_properties(self):
         for prop in ['aspect', 'facecolor', 'position', 'title',
                      'xlabel', 'xlim', 'xscale', 'xticklabels', 'xticks',
-                     'ylabel', 'ylim', 'yscale', 'yticklabels', 'yticks']:
+                     'ylabel', 'ylim', 'yscale', 'yticklabels', 'yticks',
+                     'xticklabelangle', 'yticklabelangle',
+                     ]:
             self.assertTrue(hasattr(self.axis_properties, prop))
 
     def test_has_apply_properties_method(self):
@@ -1823,6 +1927,24 @@ class TestAxisProperties(unittest.TestCase):
         with warnings.catch_warnings(record=True) as warning:
             self.axis_properties.apply(axes=plot.axes)
             self.assertFalse(len(warning))
+        plt.close(plot.figure)
+
+    def test_set_xtick_label_angle(self):
+        self.axis_properties.xticklabelangle = 45.0
+        plot = plotting.Plotter()
+        plot.plot()
+        self.axis_properties.apply(axes=plot.axes)
+        self.assertEqual(self.axis_properties.xticklabelangle,
+                         plot.axes.get_xticklabels()[0].get_rotation())
+        plt.close(plot.figure)
+
+    def test_set_ytick_label_angle(self):
+        self.axis_properties.yticklabelangle = 45.0
+        plot = plotting.Plotter()
+        plot.plot()
+        self.axis_properties.apply(axes=plot.axes)
+        self.assertEqual(self.axis_properties.yticklabelangle,
+                         plot.axes.get_yticklabels()[0].get_rotation())
         plt.close(plot.figure)